# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FOODORO POS & Kitchen Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /security/sessions:
    get:
      operationId: listSecuritySessions
      tags: [security]
      summary: List active sessions for current user (admin can pass ?userId=N)
      parameters:
        - name: userId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSession"
  /security/sessions/{id}:
    delete:
      operationId: revokeSession
      tags: [security]
      summary: Revoke a session by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Session revoked
        "403":
          description: Forbidden
        "404":
          description: Session not found
  /security/events:
    get:
      operationId: listSecurityEvents
      tags: [security]
      summary: List security events (admin/owner only)
      parameters:
        - name: type
          in: query
          schema: { type: string }
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: Security events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityEvent"
...
  /security/mfa/verify:
    post:
      operationId: verifyMfa
      tags: [security]
      summary: Confirm TOTP code to activate MFA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
      responses:
        "200":
          description: MFA activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
  /security/mfa:
    delete:
      operationId: disableMfa
      tags: [security]
      summary: Disable MFA for requesting user (admin can disable for any user via ?userId=N)
      parameters:
        - name: userId
          in: query
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
      responses:
        "200":
          description: MFA disabled

# ─── Schemas ──────────────────────────────────────────────────────────────────
components:
  schemas:
...
    UserSession:
      type: object
      required: [id, userId, isSuccess, revoked, mfaVerified, isCurrent, createdAt, lastActiveAt]
      properties:
        id:
          type: integer
        tenantId:
          type: ["integer", "null"]
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        userRole:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        userAgent:
          type: ["string", "null"]
        deviceFingerprint:
          type: ["string", "null"]
        isSuccess:
          type: boolean
        mfaVerified:
          type: boolean
        revoked:
          type: boolean
        isCurrent:
          type: boolean
        lastActiveAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    SecurityEvent:
      type: object
      required: [id, type, severity, resolved, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [login_failed, brute_force, suspicious_ip, account_locked, mfa_failed, session_revoked, mfa_enabled, mfa_disabled, login_success]
        severity:
          type: string
          enum: [low, medium, high, critical]
        ipAddress:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        metadata:
          type: ["object", "null"]
          additionalProperties: true
        resolved:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT